+++ aoc_2016/day01/part2.cpp
@@ -12,7 +12,7 @@ int Part2::solve(const vector<pair<char, int>> &moves) {
     for(const auto &move : moves) {
         auto facing_itr = find(DIRECTIONS.begin(), DIRECTIONS.end(), facing);
         int facing_index = (int) distance(DIRECTIONS.begin(), facing_itr);
-        facing = DIRECTIONS[(facing_index + (move.first == 'R' ? 1 : -1) + 4) % 4];
+        facing = DIRECTIONS[static_cast<size_t>((facing_index + (move.first == 'R' ? 1 : -1) + 4) % 4)];
         for (int i = 0; i < move.second; ++i) {
             // process all points in the move 1 by 1 to add them in the "seen" vector
             seen.insert(position);
